requirements.txt
feedparser
yt-dlp
opencv-python
pytesseract
Pillow
flask
requests
python-dotenv
schedule
